Moved PhraseTable class to `translate.api`
